+++ tests/fhir/tabulation/test_fhir_tabulation.py
@@ -24,20 +24,6 @@ from phdi.fhir.tabulation.tables import (
     extract_data_from_schema,
     _merge_include_query_params_for_location,
 )
-from phdi.fhir.utils import apply_selection_criteria
-
-
-def test_apply_selection_criteria():
-    selection_criteria_test_list = ["one", "two", "three"]
-    assert apply_selection_criteria(selection_criteria_test_list, "first") == "one"
-    assert apply_selection_criteria(selection_criteria_test_list, "last") == "three"
-    assert (
-        apply_selection_criteria(selection_criteria_test_list, "random")
-        in selection_criteria_test_list
-    )
-    assert apply_selection_criteria(selection_criteria_test_list, "all") == ",".join(
-        selection_criteria_test_list
-    )
 
 
 def test_tabulate_data_invalid_table_name():
