+++ tests/fhir/test_utils.py
@@ -3,10 +3,27 @@ import pathlib
 import pytest
 
 from phdi.fhir.utils import (
+    apply_selection_criteria,
     find_entries_by_resource_type,
     get_field,
     get_one_line_address,
 )
+
+
+def test_apply_selection_criteria():
+    with pytest.raises(ValueError) as e:
+        apply_selection_criteria([], "invalid")
+        assert (
+            'is not a valid option. Must be one of "first", "last", or "random"'
+            in str(e.value)
+        )
+    selection_criteria_test_list = ["one", "two", "three"]
+    assert apply_selection_criteria(selection_criteria_test_list, "first") == "one"
+    assert apply_selection_criteria(selection_criteria_test_list, "last") == "three"
+    assert (
+        apply_selection_criteria(selection_criteria_test_list, "random")
+        in selection_criteria_test_list
+    )
 
 
 def test_find_resource_by_type():
